AdvancedConvertor recursive
